Fix source-directories misinterpretation
When source-directories start with ".." (e.g. "../lib"), this code would
misinterpret the leading "." as the current directory and strip it off,
leaving a path that starts with "./" (which is wrong).

In short, "../lib" is interpreted as "./lib", causing diagnostics to be
attributed to a non-existent file path.

This fix simply removes the offending branch, which seems no longer to
be needed. The [original implementation][original] assembled the path
through string concatenation, but this is no longer the case.
In the current interpretation, `path.join` handles the relative paths
correctly on its own, so we can just remove this special case.

[original]: https://github.com/elm-tooling/elm-language-server/blob/85aa6a4613bd5c5adf78eb9163efb5eab93ca1b1/src/providers/diagnostics/elmMakeDiagnostics.ts#L178-L180