Send correct percentage completion values
The math was wrong on the calculation of `progressSteps`. Should have
been 100 (%) divided by the number of files to process (times two, since
there are two processing stages for each file). Before this it was the
number of files (times two) dividied by 100.

The `report` messages were also sent with the incremental deltas, but
the language server client API requires that the updated total be sent.