Removes 'Unused Value Constructor' errors on modules that expose all of their contents
Before this commit, elmLsDiagnostics doesn't present the 'Unused Value Constructor' warning for modules that either use the Value somewhere(correct) or expose the parent type explicitly with the '(..)' operator to expose the variants(this is speculative because the type may not be used elsewhere even though it's exported). This change just expands that logic to also include cases where the module implicitly exposes the parent type and its variants, by using the '(..)' indicator to expose all contents of the module. Ideally this and the other speculative case should be removed in favor of a solution that considers whether the variants are actually used in other files. The tests have been modified to prepare for such a solution and a test has been added to test the multiple file variant usage case, but in the meantime, that test also tests the single file 'expose all' case.