Fix AST getting out of sync with the document, when the client sends multiple content changes at once.
On a textDocument/didChange notification, if the contentChanges contain
multiple entries, each of them needs to be handled as if all previous
changes were already applied. When we don't, the calculation of the
change's byte offsets will be wrong, which can lead to a syntax tree
that produces parsing errors where the actual file does not have any
issues.
To solve this, we now apply the text of each change record to the full
file text and use the result when applying the next change of the
contentChanges array.